requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
anyio=3.7.1=pyhd8ed1ab_0
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py311h5eee18b_0
bcrypt=3.2.0=py311h5eee18b_1
blinker=1.7.0=pypi_0
boto3=1.34.27=pypi_0
botocore=1.34.27=pypi_0
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.12.12=h06a4308_0
cffi=1.16.0=py311h5eee18b_0
click=8.1.7=py311h06a4308_0
cryptography=41.0.7=py311hdda0065_0
ecdsa=0.17.0=pyhd3eb1b0_0
exceptiongroup=1.2.0=py311h06a4308_0
fastapi=0.109.0=pyhd8ed1ab_0
flask=3.0.1=pypi_0
future=0.18.3=py311h06a4308_0
gmp=6.2.1=h295c915_3
h11=0.12.0=pyhd3eb1b0_0
idna=3.4=py311h06a4308_0
itsdangerous=2.1.2=pypi_0
jinja2=3.1.3=pypi_0
jmespath=1.0.1=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
markupsafe=2.1.4=pypi_0
ncurses=6.4=h6a678d5_0
openssl=3.0.12=h7f8727e_0
passlib=1.7.4=pyhd3eb1b0_0
pathlib=1.0.1=pyhd3eb1b0_1
pip=23.3.1=py311h06a4308_0
pyasn1=0.4.8=pyhd3eb1b0_0
pycparser=2.21=pyhd3eb1b0_0
pycryptodome=3.15.0=py311h276157c_0
pydantic=1.10.12=py311h5eee18b_1
python=3.11.7=h955ad1f_0
python-dateutil=2.8.2=pypi_0
python-dotenv=1.0.1=pyhd8ed1ab_0
python-jose=3.3.0=pyhd3eb1b0_0
python_abi=3.11=2_cp311
readline=8.2=h5eee18b_0
rsa=4.7.2=pyhd3eb1b0_1
s3transfer=0.10.0=pypi_0
setuptools=68.2.2=py311h06a4308_0
six=1.16.0=pyhd3eb1b0_1
sniffio=1.3.0=py311h06a4308_0
sqlite=3.41.2=h5eee18b_0
starlette=0.35.0=pyhd8ed1ab_0
tk=8.6.12=h1ccaba5_0
typing-extensions=4.9.0=py311h06a4308_1
typing_extensions=4.9.0=py311h06a4308_1
tzdata=2023d=h04d1e81_0
urllib3=2.0.7=pypi_0
uvicorn=0.27.0=py311h38be061_0
werkzeug=3.0.1=pypi_0
wheel=0.41.2=py311h06a4308_0
xz=5.4.5=h5eee18b_0
zlib=1.2.13=h5eee18b_0
